+++ packages/shared-metrics/src/index.js
@@ -19,3 +19,9 @@ exports.allMetrics = [
   require('./name'),
   require('./version')
 ];
+
+exports.util = require('./util');
+
+exports.setLogger = function(logger) {
+  exports.util.setLogger(logger);
+};