+++ packages/shared-metrics/src/libuv.js
@@ -1,26 +1,32 @@
 'use strict';
 
-let logger = require('@instana/core').logger.getLogger('metrics');
-exports.setLogger = function setLogger(_logger) {
-  logger = _logger;
-};
+const path = require('path');
+
+const lag = require('event-loop-lag')(1000);
 
 let eventLoopStats;
-try {
-  eventLoopStats = require('event-loop-stats');
-} catch (e) {
-  logger.info(
-    'Could not load event-loop-stats. You will only see limited event loop information in ' +
-      'Instana for this application. This typically occurs when native addons could not be ' +
-      'installed during module installation (npm install). See the instructions to learn more ' +
-      'about the requirements of the collector: ' +
-      'https://www.instana.com/docs/ecosystem/node-js/installation/#native-addons'
-  );
-}
-const lag = require('event-loop-lag')(1000);
 
 exports.payloadPrefix = 'libuv';
 exports.currentPayload = {};
+
+const nativeModuleLoader = require('./util/nativeModuleRetry')({
+  nativeModuleName: 'event-loop-stats',
+  moduleRoot: path.join(__dirname, '..'),
+  message:
+    'Could not load event-loop-stats. You will only see limited event loop information in ' +
+    'Instana for this application. This typically occurs when native addons could not be ' +
+    'installed during module installation (npm install). See the instructions to learn more ' +
+    'about the requirements of the collector: ' +
+    'https://www.instana.com/docs/ecosystem/node-js/installation/#native-addons'
+});
+
+nativeModuleLoader.on('loaded', eventLoopStats_ => {
+  eventLoopStats = eventLoopStats_;
+});
+
+nativeModuleLoader.on('failed', () => {
+  exports.currentPayload.statsSupported = false;
+});
 
 Object.defineProperty(exports, 'currentPayload', {
   get: function() {
@@ -36,7 +42,5 @@ function sense() {
     stats.statsSupported = true;
     return stats;
   }
-  return {
-    statsSupported: false
-  };
+  return {};
 }